scripts/requirements.txt
ezdxf>=1.0.0
pdf2image>=1.16.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
